requirements.txt
selenium
allure-behave

